refactor> extract function call stuff to separate file
I saw the the official Boot.dev solution does this and to some degree it makes sense to me, decluttering the main fila is nice. Also I guess there-s some locality of behavior since the purpose of the new module is to gather what-s needed to communicate the LLM what functions it can call.